refactor: started using nuxt for vue production mode and composability
So far we imported vue development files into the browser and them ran our application in the browser. That had a few potential problems: the application could behave differently from vue recommendations, probably slower; and the browser needed to parse and evaluate all of vue code.

Most importantly though we couldn't use Single-File Components (SFC) to compose our application in a better way.

With this commit, I made the first step. I installed and configured nuxt, converted the existing pages to SFC.

I also installed prettier and configured it to format the pages directory.

I needed to upgrade vercel tool because it would not run with the newer nuxt.